+++ Code/Web Development/eCoachingLog/eCoachingLog/Services/ISurveyService.cs
@@ -4,7 +4,7 @@ namespace eCoachingLog.Services
 {
 	public interface ISurveyService
 	{
-		Survey GetSurvey(int surveyId);
+		Survey GetSurvey(long surveyId);
 		void Save(Survey survey, out int retCode, out string retMsg);
 	}
 }
